+++ include/beast/core/detail/type_traits.hpp
@@ -8,6 +8,10 @@
 #ifndef BEAST_DETAIL_TYPE_TRAITS_HPP
 #define BEAST_DETAIL_TYPE_TRAITS_HPP
 
+#include <beast/core/error.hpp>
+#include <boost/asio/buffer.hpp>
+#include <boost/asio/io_service.hpp>
+#include <iterator>
 #include <tuple>
 #include <type_traits>
 #include <stdexcept>
@@ -15,6 +19,10 @@
 
 namespace beast {
 namespace detail {
+
+//
+// utilities
+//
 
 template<class... Ts>
 struct make_void
@@ -158,6 +166,288 @@ struct get_lowest_layer
 {
     using type = typename maybe_get_lowest_layer<T,
         has_lowest_layer<T>::value>::type;
+};
+
+//------------------------------------------------------------------------------
+
+//
+// buffer concepts
+//
+
+// Types that meet the requirements,
+// for use with std::declval only.
+template<class BufferType>
+struct BufferSequence
+{
+    using value_type = BufferType;
+    using const_iterator = BufferType const*;
+    ~BufferSequence();
+    BufferSequence(BufferSequence const&) = default;
+    const_iterator begin() const noexcept;
+    const_iterator end() const noexcept;
+};
+using ConstBufferSequence =
+    BufferSequence<boost::asio::const_buffer>;
+using MutableBufferSequence =
+    BufferSequence<boost::asio::mutable_buffer>;
+
+template<class T, class BufferType>
+class is_buffer_sequence
+{
+    template<class U, class R = std::is_convertible<
+        typename U::value_type, BufferType> >
+    static R check1(int);
+    template<class>
+    static std::false_type check1(...);
+    using type1 = decltype(check1<T>(0));
+
+    template<class U, class R = std::is_base_of<
+    #if 0
+        std::bidirectional_iterator_tag,
+            typename std::iterator_traits<
+                typename U::const_iterator>::iterator_category>>
+    #else
+        // workaround:
+        // boost::asio::detail::consuming_buffers::const_iterator
+        // is not bidirectional
+        std::forward_iterator_tag,
+            typename std::iterator_traits<
+                typename U::const_iterator>::iterator_category>>
+    #endif
+    static R check2(int);
+    template<class>
+    static std::false_type check2(...);
+    using type2 = decltype(check2<T>(0));
+
+    template<class U, class R = typename
+        std::is_convertible<decltype(
+            std::declval<U>().begin()),
+                typename U::const_iterator>::type>
+    static R check3(int);
+    template<class>
+    static std::false_type check3(...);
+    using type3 = decltype(check3<T>(0));
+
+    template<class U, class R = typename std::is_convertible<decltype(
+        std::declval<U>().end()),
+            typename U::const_iterator>::type>
+    static R check4(int);
+    template<class>
+    static std::false_type check4(...);
+    using type4 = decltype(check4<T>(0));
+
+public:
+    using type = std::integral_constant<bool,
+        std::is_copy_constructible<T>::value &&
+        std::is_destructible<T>::value &&
+        type1::value && type2::value &&
+        type3::value && type4::value>;
+};
+
+template<class B1, class... Bn>
+struct is_all_ConstBufferSequence
+    : std::integral_constant<bool,
+        is_buffer_sequence<B1, boost::asio::const_buffer>::type::value &&
+        is_all_ConstBufferSequence<Bn...>::value>
+{
+};
+
+template<class B1>
+struct is_all_ConstBufferSequence<B1>
+    : is_buffer_sequence<B1, boost::asio::const_buffer>::type
+{
+};
+
+template<class T>
+class is_dynamic_buffer
+{
+    // size()
+    template<class U, class R = std::is_convertible<decltype(
+        std::declval<U const>().size()), std::size_t>>
+    static R check1(int);
+    template<class>
+    static std::false_type check1(...);
+    using type1 = decltype(check1<T>(0));
+
+    // max_size()
+    template<class U, class R = std::is_convertible<decltype(
+        std::declval<U const>().max_size()), std::size_t>>
+    static R check2(int);
+    template<class>
+    static std::false_type check2(...);
+    using type2 = decltype(check2<T>(0));
+
+    // capacity()
+    template<class U, class R = std::is_convertible<decltype(
+        std::declval<U const>().capacity()), std::size_t>>
+    static R check3(int);
+    template<class>
+    static std::false_type check3(...);
+    using type3 = decltype(check3<T>(0));
+
+    // data()
+    template<class U, class R = std::integral_constant<
+        bool, is_buffer_sequence<decltype(
+            std::declval<U const>().data()),
+                boost::asio::const_buffer>::type::value>>
+    static R check4(int);
+    template<class>
+    static std::false_type check4(...);
+    using type4 = decltype(check4<T>(0));
+
+    // prepare()
+    template<class U, class R = std::integral_constant<
+        bool, is_buffer_sequence<decltype(
+            std::declval<U>().prepare(1)),
+                boost::asio::mutable_buffer>::type::value>>
+    static R check5(int);
+    template<class>
+    static std::false_type check5(...);
+    using type5 = decltype(check5<T>(0));
+
+    // commit()
+    template<class U, class R = decltype(
+        std::declval<U>().commit(1), std::true_type{})>
+    static R check6(int);
+    template<class>
+    static std::false_type check6(...);
+    using type6 = decltype(check6<T>(0));
+
+    // consume
+    template<class U, class R = decltype(
+        std::declval<U>().consume(1), std::true_type{})>
+    static R check7(int);
+    template<class>
+    static std::false_type check7(...);
+    using type7 = decltype(check7<T>(0));
+
+public:
+    using type = std::integral_constant<bool,
+        type1::value
+        && type2::value
+        //&& type3::value // Networking TS
+        && type4::value
+        && type5::value
+        && type6::value
+        && type7::value
+    >;
+};
+
+//------------------------------------------------------------------------------
+
+//
+// stream concepts
+//
+
+// Types that meet the requirements,
+// for use with std::declval only.
+struct StreamHandler
+{
+    StreamHandler(StreamHandler const&) = default;
+    void operator()(error_code ec, std::size_t);
+};
+using ReadHandler = StreamHandler;
+using WriteHandler = StreamHandler;
+
+template<class T>
+class has_get_io_service
+{
+    template<class U, class R = typename std::is_same<
+        decltype(std::declval<U>().get_io_service()),
+            boost::asio::io_service&>>
+    static R check(int);
+    template<class>
+    static std::false_type check(...);
+public:
+    using type = decltype(check<T>(0));
+};
+
+template<class T>
+class is_async_read_stream
+{
+    template<class U, class R = decltype(
+        std::declval<U>().async_read_some(
+            std::declval<MutableBufferSequence>(),
+                std::declval<ReadHandler>()),
+                    std::true_type{})>
+    static R check(int);
+    template<class>
+    static std::false_type check(...);
+    using type1 = decltype(check<T>(0));
+public:
+    using type = std::integral_constant<bool,
+        type1::value &&
+        has_get_io_service<T>::type::value>;
+};
+
+template<class T>
+class is_async_write_stream
+{
+    template<class U, class R = decltype(
+        std::declval<U>().async_write_some(
+            std::declval<ConstBufferSequence>(),
+                std::declval<WriteHandler>()),
+                    std::true_type{})>
+    static R check(int);
+    template<class>
+    static std::false_type check(...);
+    using type1 = decltype(check<T>(0));
+public:
+    using type = std::integral_constant<bool,
+        type1::value &&
+        has_get_io_service<T>::type::value>;
+};
+
+template<class T>
+class is_sync_read_stream
+{
+    template<class U, class R = std::is_same<decltype(
+        std::declval<U>().read_some(
+            std::declval<MutableBufferSequence>())),
+                std::size_t>>
+    static R check1(int);
+    template<class>
+    static std::false_type check1(...);
+    using type1 = decltype(check1<T>(0));
+
+    template<class U, class R = std::is_same<decltype(
+        std::declval<U>().read_some(
+            std::declval<MutableBufferSequence>(),
+                std::declval<error_code&>())), std::size_t>>
+    static R check2(int);
+    template<class>
+    static std::false_type check2(...);
+    using type2 = decltype(check2<T>(0));
+
+public:
+    using type = std::integral_constant<bool,
+        type1::value && type2::value>;
+};
+
+template<class T>
+class is_sync_write_stream
+{
+    template<class U, class R = std::is_same<decltype(
+        std::declval<U>().write_some(
+            std::declval<ConstBufferSequence>())),
+                std::size_t>>
+    static R check1(int);
+    template<class>
+    static std::false_type check1(...);
+    using type1 = decltype(check1<T>(0));
+
+    template<class U, class R = std::is_same<decltype(
+        std::declval<U>().write_some(
+            std::declval<ConstBufferSequence>(),
+                std::declval<error_code&>())), std::size_t>>
+    static R check2(int);
+    template<class>
+    static std::false_type check2(...);
+    using type2 = decltype(check2<T>(0));
+
+public:
+    using type = std::integral_constant<bool,
+        type1::value && type2::value>;
 };
 
 } // detail