+++ test/core/multi_buffer.cpp
@@ -9,8 +9,8 @@
 #include <beast/core/multi_buffer.hpp>
 
 #include "buffer_test.hpp"
-#include <beast/core/buffer_concepts.hpp>
 #include <beast/core/ostream.hpp>
+#include <beast/core/type_traits.hpp>
 #include <beast/test/test_allocator.hpp>
 #include <beast/unit_test/suite.hpp>
 #include <boost/asio/buffer.hpp>